Move from Users namespace to user
