+++ src/Umbraco.Cms.Api.Management/Factories/IUserPresentationFactory.cs
@@ -1,4 +1,4 @@
-﻿using Umbraco.Cms.Api.Management.ViewModels.Users;
+﻿using Umbraco.Cms.Api.Management.ViewModels.User;
 using Umbraco.Cms.Core.Models;
 using Umbraco.Cms.Core.Models.Membership;
