+++ src/Umbraco.Cms.Api.Management/Mapping/Items/ItemTypeMapDefinition.cs
@@ -9,8 +9,8 @@ using Umbraco.Cms.Api.Management.ViewModels.MemberGroup.Item;
 using Umbraco.Cms.Api.Management.ViewModels.MemberType.Items;
 using Umbraco.Cms.Api.Management.ViewModels.RelationType.Item;
 using Umbraco.Cms.Api.Management.ViewModels.Template.Item;
+using Umbraco.Cms.Api.Management.ViewModels.User.Item;
 using Umbraco.Cms.Api.Management.ViewModels.UserGroups.Item;
-using Umbraco.Cms.Api.Management.ViewModels.Users.Item;
 using Umbraco.Cms.Core.Mapping;
 using Umbraco.Cms.Core.Models;
 using Umbraco.Cms.Core.Models.Entities;
