+++ src/Umbraco.Cms.Api.Management/ViewModels/User/Item/UserItemResponseModel.cs
@@ -1,6 +1,6 @@
 ﻿using Umbraco.Cms.Api.Management.ViewModels.Item;
 
-namespace Umbraco.Cms.Api.Management.ViewModels.Users.Item;
+namespace Umbraco.Cms.Api.Management.ViewModels.User.Item;
 
 public class UserItemResponseModel : ItemResponseModelBase
 {
