+++ src/Umbraco.Cms.Api.Management/ViewModels/User/UserResponseModel.cs
@@ -1,6 +1,6 @@
 ﻿using Umbraco.Cms.Core.Models.Membership;
 
-namespace Umbraco.Cms.Api.Management.ViewModels.Users;
+namespace Umbraco.Cms.Api.Management.ViewModels.User;
 
 public class UserResponseModel : UserPresentationBase, INamedEntityPresentationModel
 {
